TDL-19620 Refactoring changes (#6)
* refactoring for exception classes & helper methods

* directory struct change:improving readability for streams.py

* removed abstract decorator for set_params method

* Updated Circle Ci configurtion file

* Updated the formatting of the function call as per the suggestion


* formatting changes
