+++ tap_dixa/__init__.py
@@ -1,6 +1,5 @@
 import singer
 from singer import utils
-
 from tap_dixa.discover import discover
 from tap_dixa.sync import sync
 
@@ -15,14 +14,14 @@ def main():
 
     # If discover flag was passed, run discovery mode and dump output to stdout
     if args.discover:
-        catalog = discover()
+        catalog = discover(args.config)
         catalog.dump()
     # Otherwise run in sync mode
     else:
         if args.catalog:
             catalog = args.catalog
         else:
-            catalog = discover()
+            catalog = discover(args.config)
         sync(args.config, args.state, catalog)
