+++ tap_dixa/helpers.py
@@ -1,18 +1,23 @@
+""" helper methods required for tap-dixa"""
 import datetime
+import os
+from enum import Enum
 from typing import Iterator
 from urllib.parse import parse_qsl, urlparse
 
-import singer
+from singer import get_logger, utils
+
+LOGGER = get_logger()
 
 
-def unix_ms_to_date(ms: int) -> str:
+def unix_ms_to_date(timestamp_ms: int) -> str:
     """
     Converts unix timestamp in milliseconds to ISO 8601 date string.
 
     :param ms: unix timestamp in milliseconds
     :return: ISO 8601 date string
     """
-    return datetime.datetime.fromtimestamp(ms/1000).replace(microsecond=0).isoformat()
+    return datetime.datetime.fromtimestamp(timestamp_ms / 1000).replace(microsecond=0).isoformat()
 
 
 def datetime_to_unix_ms(datetime_obj: datetime.datetime) -> int:
@@ -36,9 +41,7 @@ def create_csid_params(csids: Iterator) -> dict:
     :param csids: Iterator containing conversation IDs
     :return: dictionary to be used with activity logs endpoint
     """
-    return {
-        'csids': ','.join(map(str, csids))
-    }
+    return {"csids": ",".join(map(str, csids))}
 
 
 def chunks(arr, chunk_size=10):
@@ -51,14 +54,14 @@ def chunks(arr, chunk_size=10):
     :return: a chunk sized array
     """
     for i in range(0, len(arr), chunk_size):
-        yield arr[i:i + chunk_size]
+        yield arr[i: i + chunk_size]
 
 
 def date_to_rfc3339(date: str) -> str:
     """Converts date to rfc 3339"""
-    d = singer.utils.strptime_to_utc(date)
+    date_utc = utils.strptime_to_utc(date)
 
-    return d.strftime('%Y-%m-%dT%H:%M:%SZ')
+    return date_utc.strftime("%Y-%m-%dT%H:%M:%SZ")
 
 
 def get_next_page_key(url_part: str) -> dict:
@@ -69,3 +72,53 @@ def get_next_page_key(url_part: str) -> dict:
     parsed_url = urlparse(url_part)
 
     return dict(parse_qsl(parsed_url.query))
+
+
+def get_abs_path(path):
+    """
+    Gets the absolute path of the provided relative path.
+    """
+    return os.path.join(os.path.dirname(os.path.realpath(__file__)), path)
+
+
+def _get_key_properties_from_meta(schema_meta):
+    """
+    Retrieves the 'table-key-properties' value from the schema metadata.
+    """
+    return schema_meta[0].get("metadata").get("table-key-properties")
+
+
+def _get_replication_method_from_meta(schema_meta):
+    """
+    Retrieves the 'forced-replication-method' value from the schema metadata.
+    """
+    return schema_meta[0].get("metadata").get("forced-replication-method")
+
+
+def _get_replication_key_from_meta(schema_meta):
+    """
+    Retrieves the 'valid-replication-keys' value from the schema metadata.
+    """
+    if _get_replication_method_from_meta(schema_meta) == "INCREMENTAL":
+        return schema_meta[0].get("metadata").get("valid-replication-keys")[0]
+    return None
+
+
+class Interval(Enum):
+    """
+    Enum representing time interval for making API calls.
+    """
+
+    HOUR = 1
+    DAY = 24
+    WEEK = 24 * 7
+    MONTH = 24 * 31
+
+
+class DixaURL(Enum):
+    """
+    Enum representing the Dixa base url API variants.
+    """
+
+    EXPORTS = "https://exports.dixa.io"
+    INTEGRATIONS = "https://dev.dixa.io"
