--- tap_dixa/schemas/al.json
@@ -1,5 +1,0 @@
-{
-  "type": [
-    "null"
-  ]
-}
