+++ tap_dixa/streams/__init__.py
@@ -1,0 +1,9 @@
+from .activitylogs import ActivityLogs
+from .conversations import Conversations
+from .messages import Messages
+
+STREAMS = {
+    "activity_logs": ActivityLogs,
+    "conversations": Conversations,
+    "messages": Messages,
+}
